Now multiple types of submission can be displayed in the same table
Moved the paginationHelper into the submissionHelper because it made
more sense.
Added the ability to excludeTypes from a submission query.  So we get
all of the submissions and filter out the ones that we don't want on the
server.
Added a new json file because the tables are diverging and it make more
sense to use different techniques on each.